+++ packages/react-dom/tests/Link.spec.tsx
@@ -307,7 +307,7 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
     });
   });
 
-  describe("anchorProps", () => {
+  describe("forward", () => {
     describe("additional props (deprecated)", () => {
       it("warns when passing additional props to the <Link>", () => {
         const realWarn = console.warn;
@@ -329,9 +329,9 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
         expect(fakeWarn.mock.calls[0][0]).toBe(`Deprecation warning:
 Passing additional props to a <Link> will no longer be forwarded to the rendered component in v2.
 
-Instead, please use the "anchorProps" prop to pass an object of props to be attached to the component.
+Instead, please use the "forward" prop to pass an object of props to be attached to the component.
 
-<Link to="Route Name" anchorProps={{ className: "test" }}>`);
+<Link to="Route Name" forward={{ className: "test" }}>`);
 
         console.warn = realWarn;
       });
@@ -353,11 +353,11 @@ Instead, please use the "anchorProps" prop to pass an object of props to be atta
       });
     });
 
-    it("passes anchorProps to the rendered anchor", () => {
+    it("passes forward to the rendered anchor", () => {
       ReactDOM.render(
         <Router>
           {() => (
-            <Link to="Test" anchorProps={{ className: "hi" }}>
+            <Link to="Test" forward={{ className: "hi" }}>
               Test
             </Link>
           )}