+++ packages/react-native/src/Link.tsx
@@ -22,7 +22,7 @@ export interface LinkProps {
   style?: any;
   method?: NavType;
   children: NavigatingChildren | React.ReactNode;
-  anchorProps?: object;
+  forward?: object;
 }
 
 interface BaseLinkProps extends LinkProps {
@@ -35,7 +35,7 @@ interface LinkState {
 }
 
 let hasWarnedTo = false;
-let hasWarnedAnchorProps = false;
+let hasWarnedForward = false;
 
 class BaseLink extends React.Component<BaseLinkProps, LinkState> {
   removed: boolean;
@@ -106,7 +106,7 @@ class BaseLink extends React.Component<BaseLinkProps, LinkState> {
       method,
       forwardedRef,
       children,
-      anchorProps,
+      forward,
       ...rest
     } = this.props;
     if (process.env.NODE_ENV !== "production") {
@@ -118,14 +118,14 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
 <Link name="Route Name">...</Link>`);
       }
 
-      if (!hasWarnedAnchorProps && Object.keys(rest).length > 0) {
-        hasWarnedAnchorProps = true;
+      if (!hasWarnedForward && Object.keys(rest).length > 0) {
+        hasWarnedForward = true;
         console.warn(`Deprecation warning:
 Passing additional props to a <Link> will no longer be forwarded to the rendered component in v2.
 
-Instead, please use the "anchorProps" prop to pass an object of props to be attached to the component.
+Instead, please use the "forward" prop to pass an object of props to be attached to the component.
 
-<Link to="Route Name" anchorProps={{ className: "test" }}>`);
+<Link to="Route Name" forward={{ className: "test" }}>`);
       }
     }
     
@@ -133,7 +133,7 @@ Instead, please use the "anchorProps" prop to pass an object of props to be atta
 
     const additionalProps = {
       ...rest,
-      ...anchorProps
+      ...forward
     };
 
     return (