+++ packages/react-native/tests/Link.spec.tsx
@@ -225,7 +225,7 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
     });
   });
 
-  describe("anchorProps", () => {
+  describe("forward", () => {
     describe("additional props (deprecated)", () => {
       it("warns when passing additional props to the <Link>", () => {
         const realWarn = console.warn;
@@ -257,9 +257,9 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
         expect(fakeWarn.mock.calls[0][0]).toBe(`Deprecation warning:
 Passing additional props to a <Link> will no longer be forwarded to the rendered component in v2.
 
-Instead, please use the "anchorProps" prop to pass an object of props to be attached to the component.
+Instead, please use the "forward" prop to pass an object of props to be attached to the component.
 
-<Link to="Route Name" anchorProps={{ className: "test" }}>`);
+<Link to="Route Name" forward={{ className: "test" }}>`);
 
         console.warn = realWarn;
       });
@@ -290,7 +290,7 @@ Instead, please use the "anchorProps" prop to pass an object of props to be atta
       });
     });
 
-    it("passes anchorProps to the rendered anchor", () => {
+    it("passes forward to the rendered anchor", () => {
       const history = InMemory({ locations: ["/the-initial-location"] });
       const mockNavigate = jest.fn();
       history.navigate = mockNavigate;
@@ -305,7 +305,7 @@ Instead, please use the "anchorProps" prop to pass an object of props to be atta
       const tree = renderer.create(
         <Router>
           {() => (
-            <Link to="Test" anchorProps={{ style }}>
+            <Link to="Test" forward={{ style }}>
               <Text>Test</Text>
             </Link>
           )}