+++ packages/react-native/types/Link.d.ts
@@ -15,7 +15,7 @@ export interface LinkProps {
     style?: any;
     method?: NavType;
     children: NavigatingChildren | React.ReactNode;
-    anchorProps?: object;
+    forward?: object;
 }
 declare const Link: React.ComponentType<LinkProps & React.ClassAttributes<{}>>;
 export default Link;