+++ packages/react-native/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next
 
-* Add `anchorProps` prop to `Link` for passing props to rendered component.
+* Add `forward` prop to `Link` for passing props to rendered component.
 * Deprecate passing props from `Link` to rendered component.
 * `<Link>` takes `name` prop, preferable over `to`.
 * Only export public TypeScript types.